backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
sendgrid==6.11.0
email-validator==2.1.0
python-dotenv==1.0.0
qrcode[pil]==7.4.2
pillow==10.1.0
pandas==2.1.4
openpyxl==3.1.2
scikit-learn==1.3.2
numpy==1.24.4
websockets==12.0
httpx==0.25.2
reportlab==4.0.8
pdfdocument==4.0.0
matplotlib==3.8.2
seaborn==0.13.0
playwright
